Adapted gnu_linux_dpdk to the new bufferpool
