+++ src/xdpd/drivers/gnu_linux/src/io/bufferpool.cc
@@ -70,41 +70,3 @@ bufferpool::~bufferpool(){
 		free(pool[i].pkt);
 	}
 }
-
-
-
-//
-// Buffer pool management
-//
-void bufferpool::init(){
-	
-	pthread_mutex_lock(&bufferpool::mutex);		
-
-	if(bufferpool::instance){
-		//Double-call to init??
-		ROFL_DEBUG(DRIVER_NAME"[bufferpool] Double call to bufferpool init!! Skipping...\n");
-		pthread_mutex_unlock(&bufferpool::mutex);
-		return;	
-	}
-	
-	ROFL_DEBUG(DRIVER_NAME"[bufferpool] Initializing bufferpool with a capacity of %d buffers...\n",capacity);
-
-	//Init 	
-	bufferpool::instance = new bufferpool();
-	
-	ROFL_DEBUG(DRIVER_NAME"[bufferpool] Initialization was successful\n");
-
-	//Wake consumers
-	pthread_cond_broadcast(&bufferpool::cond);
-
-	//Release and go!	
-	pthread_mutex_unlock(&bufferpool::mutex);		
-}
-
-void bufferpool::destroy(){
-
-	if(get_instance())
-		delete get_instance();
-
-	instance = NULL;	
-}
