+++ src/xdpd/drivers/gnu_linux_dpdk/src/hal-imp/driver.cc
@@ -107,7 +107,7 @@ hal_result_t hal_driver_init(hal_extension_ops_t* extensions, const char* extra_
 #endif
 
 	//Init bufferpool
-	bufferpool::init(IO_BUFFERPOOL_RESERVOIR);
+	bufferpool::init();
 
 	//Initialize pipeline
 	if(physical_switch_init() != ROFL_SUCCESS)
